Initial version of working version - image upload
Here we have a working example based on the example found
at: https://blog.miguelgrinberg.com/post/handling-file-uploads-with-flask

After few modifications such as the removal of deprecated module
`imghdr` we were able to make it work properly with other small
adjustments.

Now the next step is to integrate it with database.